Created function that creates array of GifCard components
In order to populate page with gifs we need a function to extract all the arrays within the state.
A map is used to parse the array and send each individual gif's url and title to a GifCard component.
The array of the newly created gifs are then returned in the render statement.